Change AIE show buttons to use focus Split AIE functions into files
